[Fix] Add missing asPath for ServerPageContext
ServerPageContext has `asPath` in types, but actually next.js doesn't provide it, instead we have `resolvedUrl` in `GetServerSidePropsContext` with the same value as `asPath`. So, for consistency, we need to add it with `asPath` field, because, for example when you try to switch your page from gip to gssp, you may encounter an inability to retrieve this value through the page event.